requirements.txt
aniso8601==3.0.2
certifi==2018.8.24
chardet==3.0.4
click==6.7
distro==1.3.0
Flask==1.0.2
Flask-RESTful==0.3.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.1
pandas==0.23.4
PyGithub==1.40
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
tabula-py==1.2.0
urllib3==1.23
Werkzeug==0.14.1
